Move the charge logic to control pallet & More refactor
